README.md
# Discord-Tagger---Raid-Tool
This is my Discord Tagger it helps people to Raid Discord Servers and spam Discord DMs. Though this tool is made in Python I converted it to .exe for other users who don't have Python, both are included in this repository as well as my YouTube Channel.

How to use?
For Normal Users: Run the .exe file in the folder.
For Advanced users: Type `pip install pyinstaller` in your Command Prompt and run the .py file!

This is my first every project in Python and I'll introduce it's more versions too!

Happy Raiding!